structured the profile screen, separating widgets and logic.
