+++ lib/player/ui/screens/profile.screen.dart
@@ -1,38 +1,23 @@
 import 'dart:async';
-import 'dart:io';
-
-import 'package:file_picker/file_picker.dart';
-import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/material.dart';
-import 'package:lab_movil_2222/assets/image/services/upload-image.service.dart';
 import 'package:lab_movil_2222/models/asset.dto.dart';
-import 'package:lab_movil_2222/player/models/award.model.dart';
 import 'package:lab_movil_2222/player/models/player.model.dart';
 import 'package:lab_movil_2222/player/services/get-current-player.service.dart';
-import 'package:lab_movil_2222/player/services/update-avatar.service.dart';
-import 'package:lab_movil_2222/player/widgets/user-gamification.widget.dart';
+import 'package:lab_movil_2222/player/widgets/changeAvatarButton.widget.dart';
+import 'package:lab_movil_2222/player/widgets/gamification-item.widget.dart';
+import 'package:lab_movil_2222/player/widgets/player-gammification.widget.dart';
 import 'package:lab_movil_2222/points-explanation/uid/widgets/points-explanation.widget.dart';
 import 'package:lab_movil_2222/services/current-user.service.dart';
 import 'package:lab_movil_2222/shared/widgets/app-loading.widget.dart';
 import 'package:lab_movil_2222/shared/widgets/days_left_widget.dart';
-import 'package:lab_movil_2222/shared/widgets/medals-list-item_widget.dart';
 import 'package:lab_movil_2222/themes/colors.dart';
 import 'package:lab_movil_2222/user/widgets/login.screen.dart';
 import 'package:lab_movil_2222/widgets/decorated-background/background-decoration.widget.dart';
 import 'package:lab_movil_2222/widgets/scaffold-2222/bottom-navigation-bar-widget.dart';
 import 'package:lab_movil_2222/widgets/scaffold-2222/scaffold-2222.widget.dart';
-import 'package:provider/provider.dart';
-import 'package:rxdart/rxdart.dart';
 
 class ProfileScreen extends StatefulWidget {
   static const String route = '/profile';
-
-  final UpdateAvatarService _updateAvatarService;
-
-  ProfileScreen({
-    Key? key,
-  })  : this._updateAvatarService = UpdateAvatarService(),
-        super(key: key);
 
   @override
   _ProfileScreenState createState() => _ProfileScreenState();
@@ -41,9 +26,7 @@ class ProfileScreen extends StatefulWidget {
 class _ProfileScreenState extends State<ProfileScreen> {
   StreamSubscription? _signOutSub;
   StreamSubscription? _loadPlayerSub;
-  StreamSubscription? _uploadFileSub;
   PlayerModel? player;
-  ImageDto? avatarImage;
 
   @override
   void initState() {
@@ -67,12 +50,6 @@ class _ProfileScreenState extends State<ProfileScreen> {
 
   @override
   Widget build(BuildContext context) {
-    avatarImage = this.player?.avatarImage;
-    int numberProjects = this.player!.projectAwards.length;
-    int numberClubhouses = this.player!.clubhouseAwards.length;
-    int numberProactivity = numberProjects + numberClubhouses;
-    // int numberContributions = this.player!.contributionsAwards.length;
-
     final double sideSpacing = MediaQuery.of(context).size.width * 0.08;
 
     return Scaffold2222.navigation(
@@ -130,14 +107,11 @@ class _ProfileScreenState extends State<ProfileScreen> {
                               Positioned(
                                 bottom: -6,
                                 right: -14,
-                                child: ElevatedButton(
-                                  onPressed: _changeAvatar,
-                                  style: ElevatedButton.styleFrom(
-                                      shape: CircleBorder(),
-                                      primary: Colors.black,
-                                      onPrimary: Colors.white,
-                                      elevation: 0),
-                                  child: Icon(Icons.photo_camera_rounded),
+
+                                /// implements the widget to change the avatar
+                                /// logic is implemented on the button
+                                child: ChangeAvatarButton(
+                                  player: this.player!,
                                 ),
                               ),
                             ]),
@@ -176,75 +150,8 @@ class _ProfileScreenState extends State<ProfileScreen> {
                 ),
               ),
 
-              /// player's gammification
-              GammificationWidget(
-                number: numberProactivity,
-                kind: GammificationKind.proactivity,
-              ),
-
-              SizedBox(
-                height: 32,
-              ),
-              Container(
-                height: 1,
-                color: Colors2222.white.withOpacity(0.5),
-              ),
-              SizedBox(
-                height: 32,
-              ),
-
-              GammificationWidget(
-                number: numberProjects,
-                kind: GammificationKind.prizes,
-                image: AssetImage('assets/gamification/2222_monedas.png'),
-              ),
-              SizedBox(
-                height: 40,
-              ),
-
-              GammificationWidget(
-                number: numberClubhouses,
-                kind: GammificationKind.clubhouse,
-                image: AssetImage(
-                    'assets/gamification/2222_medalla-clubhouse.png'),
-              ),
-              SizedBox(
-                height: 64,
-              ),
-
-              // Padding(
-              //   padding: EdgeInsets.only(bottom: 10),
-              //   child: Text(
-              //     'Premios Obtenidos',
-              //     style: Theme.of(context).textTheme.headline5,
-              //     textAlign: TextAlign.center,
-              //   ),
-              // ),
-
-              // for (AwardModel medal in this.player!.projectAwards)
-              //   MedalsListItem(
-              //     cityName: medal.cityId,
-              //   ),
-
-              // MedalsListItem(
-              //   cityName: 'Quitu',
-              // ),
-
-              // /// project medals
-              // if (this.player!.clubhouseAwards.length > 0) ...[
-              //   Padding(
-              //     padding: EdgeInsets.only(bottom: 10),
-              //     child: Text(
-              //       'Medallas clubhouse',
-              //       style: Theme.of(context).textTheme.headline5,
-              //       textAlign: TextAlign.center,
-              //     ),
-              //   ),
-              //   for (AwardModel medal in this.player!.clubhouseAwards)
-              //     MedalsListItem(
-              //       cityName: medal.cityId,
-              //     ),
-              // ],
+              /// widget that contains a list of the player's gammification
+              PlayerGammification(player: this.player!),
 
               Padding(
                 padding: const EdgeInsets.only(bottom: 25),
@@ -321,25 +228,5 @@ class _ProfileScreenState extends State<ProfileScreen> {
         style: textTheme.headline6?.apply(),
       ),
     );
-  }
-
-  /// method to change the avatarImage
-  void _changeAvatar() async {
-    if (this._uploadFileSub != null) return;
-    String oldUrl = this.player!.avatarImage.url;
-    print('old image url');
-    UploadImageService uploadImageService =
-        Provider.of<UploadImageService>(context, listen: false);
-
-    this._uploadFileSub = uploadImageService
-        .upload$('players/${this.player!.uid}/assets')
-        .switchMap((image) =>
-            this.widget._updateAvatarService.updateAvatar$(image, oldUrl))
-        .listen((sended) {
-      if (sended) print('Imagen cambiada correctamente');
-    }, onDone: () {
-      this._uploadFileSub?.cancel();
-      this._uploadFileSub = null;
-    });
   }
 }
